creates scalar aggregation view as generalization of storybook; preparation for integration into actual frontend
