+++ src/app/patterns/body/scalars/ScalarAggregationSmallMultiple.svelte
@@ -4,25 +4,24 @@ const largestPercentile = writable(0);
 
 <script>
 import { writable, derived } from 'svelte/store';
-import { slide } from 'svelte/transition';
 
 export let data;
 export let key;
 export let resolution = 'ALL_TIME';
 export let percentiles = [50];
 
-import DataGraphic from '../../../src/components/data-graphics/DataGraphic.svelte';
-import GraphicBody from '../../../src/components/data-graphics/GraphicBody.svelte';
-import BottomAxis from '../../../src/components/data-graphics/BottomAxis.svelte';
-import LeftAxis from '../../../src/components/data-graphics/LeftAxis.svelte';
-import BuildIDRollover from '../../../src/components/data-graphics/rollovers/BuildIDRollover.svelte';
-import Line from '../../../src/components/data-graphics/LineMultiple.svelte';
-import ComparisonSummary from '../../../src/components/data-graphics/ComparisonSummary.svelte';
+import DataGraphic from '../../../../components/data-graphics/DataGraphic.svelte';
+import GraphicBody from '../../../../components/data-graphics/GraphicBody.svelte';
+import BottomAxis from '../../../../components/data-graphics/BottomAxis.svelte';
+import LeftAxis from '../../../../components/data-graphics/LeftAxis.svelte';
+import BuildIDRollover from '../../../../components/data-graphics/rollovers/BuildIDRollover.svelte';
+import Line from '../../../../components/data-graphics/LineMultiple.svelte';
+import ComparisonSummary from '../../../../components/data-graphics/ComparisonSummary.svelte';
 
 import {
   buildIDToDate, firstOfMonth, buildIDToMonth, mondays, getFirstBuildOfDays,
-} from '../../../src/components/data-graphics/utils/build-id-utils';
-import { extractPercentiles } from '../../../src/components/data-graphics/utils/percentiles';
+} from '../../../../components/data-graphics/utils/build-id-utils';
+import { extractPercentiles } from '../../../../components/data-graphics/utils/percentiles';
 
 let domain = writable(data.map((d) => d.label));
 
@@ -105,19 +104,21 @@ let latest = data[data.length - 1];
 
 <div class=graphic-and-summary>
   <DataGraphic
-    bind:margins={margins}
-    bind:rollover={dgRollover}
-    bind:xScale={xScale}
-    bind:bodyHeight={H}
-    bind:topPlot={T}
-    key={key}
     data={data}
     xDomain={$domain}
     yDomain={[0, largestPercentileValue]}
     yType="numeric"
     width=400
     height=150
+    
     bind:dataGraphicMounted={dataGraphicMounted}
+    bind:margins={margins}
+    bind:rollover={dgRollover}
+    bind:xScale={xScale}
+    bind:bodyHeight={H}
+    bind:topPlot={T}
+    key={key}
+
   >
     <LeftAxis showBorder=true />
     <BottomAxis ticks={ticks} tickFormatter={tickFormatter} />
@@ -141,7 +142,7 @@ let latest = data[data.length - 1];
         label={rollover.datum.label}
       />
       <rect x={xScale(rollover.x) - xScale.step() / 2} y={topPlot} width={xScale.step()} height={bodyHeight}
-     fill="var(--cool-gray-700)" opacity=.2 />
+      fill="var(--cool-gray-700)" opacity=.2 />
     {/if}
   </DataGraphic>
   <ComparisonSummary
